Untangle it!
